pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pr-security-review"
version = "0.1.0"
description = "Automated security review of pull requests using LLMs"
requires-python = ">=3.8"
dependencies = [
    "anthropic>=0.46.0",
    "openai>=1.63.2",
    "google-generativeai>=0.3.0",
    "PyGithub>=2.6.0",
    "packaging>=24.2",
    "numpy>=2.2.3",
    "requests>=2.32.3",
    "python-dotenv>=1.0.0",
    "markdown>=3.4.0",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=0.5.0",
    "google-auth-httplib2>=0.1.0",
    "flask>=2.3.0",
    "flask-session>=0.5.0",
    "boto3>=1.35.0",
    "gunicorn>=21.0.0",
    "pika>=1.3.0"
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0"
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-v --cov=pr_security_review --cov-report=term-missing"
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.coverage.run]
source = ["pr_security_review"]
omit = [
    "tests/*",
    "**/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "pass",
    "raise ImportError",
]
